+++ poem/src/error.rs
@@ -7,6 +7,7 @@ use std::{
     string::FromUtf8Error,
 };
 
+use headers::{ContentRange, HeaderMapExt};
 use http::Method;
 
 use crate::{http::StatusCode, IntoResponse, Response};
@@ -28,6 +29,16 @@ macro_rules! define_http_error {
 pub trait ResponseError {
     /// The status code of this error.
     fn status(&self) -> StatusCode;
+
+    /// Convert this error to a HTTP response.
+    fn as_response(&self) -> Response
+    where
+        Self: StdError + Send + Sync + 'static,
+    {
+        Response::builder()
+            .status(self.status())
+            .body(self.to_string())
+    }
 }
 
 enum ErrorSource {
@@ -42,6 +53,34 @@ impl Debug for ErrorSource {
             ErrorSource::BoxedError(err) => Debug::fmt(err, f),
             #[cfg(feature = "anyhow")]
             ErrorSource::Anyhow(err) => Debug::fmt(err, f),
+        }
+    }
+}
+
+type BoxAsResponseFn = Box<dyn Fn(&Error) -> Response + Send + Sync + 'static>;
+
+enum AsResponse {
+    Status(StatusCode),
+    Fn(BoxAsResponseFn),
+}
+
+impl AsResponse {
+    #[inline]
+    fn from_status(status: StatusCode) -> Self {
+        AsResponse::Status(status)
+    }
+
+    fn from_type<T: ResponseError + StdError + Send + Sync + 'static>() -> Self {
+        AsResponse::Fn(Box::new(|err| {
+            let err = err.downcast_ref::<T>().expect("valid error");
+            err.as_response()
+        }))
+    }
+
+    fn as_response(&self, err: &Error) -> Response {
+        match self {
+            AsResponse::Status(status) => Response::builder().status(*status).body(err.to_string()),
+            AsResponse::Fn(f) => f(err),
         }
     }
 }
@@ -62,7 +101,7 @@ impl Debug for ErrorSource {
 /// # Create you own error type
 ///
 /// ```
-/// use poem::{error::ResponseError, handler, http::StatusCode, Result};
+/// use poem::{error::ResponseError, handler, http::StatusCode, Endpoint, Request, Result};
 ///
 /// #[derive(Debug, thiserror::Error)]
 /// #[error("my error")]
@@ -75,17 +114,65 @@ impl Debug for ErrorSource {
 /// }
 ///
 /// fn do_something() -> Result<(), MyError> {
-///     todo!()
+///     Err(MyError)
 /// }
 ///
 /// #[handler]
 /// async fn index() -> Result<()> {
 ///     Ok(do_something()?)
 /// }
+///
+/// # tokio::runtime::Runtime::new().unwrap().block_on(async {
+/// let resp = index.get_response(Request::default()).await;
+/// assert_eq!(resp.status(), StatusCode::BAD_GATEWAY);
+/// assert_eq!(resp.into_body().into_string().await.unwrap(), "my error");
+/// # });
+/// ```
+///
+/// # Custom error response
+///
+/// ```
+/// use poem::{error::ResponseError, handler, http::StatusCode, Response, Result, Request, Body, Endpoint};
+///
+/// #[derive(Debug, thiserror::Error)]
+/// #[error("my error")]
+/// struct MyError;
+///
+/// impl ResponseError for MyError {
+///     fn status(&self) -> StatusCode {
+///         StatusCode::BAD_GATEWAY
+///     }
+///
+///     fn as_response(&self) -> Response {
+///         let body = Body::from_json(serde_json::json!({
+///             "code": 1000,
+///             "message": self.to_string(),
+///         })).unwrap();
+///         Response::builder().status(self.status()).body(body)
+///     }
+/// }
+///
+/// fn do_something() -> Result<(), MyError> {
+///     Err(MyError)
+/// }
+///
+/// #[handler]
+/// async fn index() -> Result<()> {
+///     Ok(do_something()?)
+/// }
+///
+/// # tokio::runtime::Runtime::new().unwrap().block_on(async {
+/// let resp = index.get_response(Request::default()).await;
+/// assert_eq!(resp.status(), StatusCode::BAD_GATEWAY);
+/// assert_eq!(resp.into_body().into_json::<serde_json::Value>().await.unwrap(),
+/// serde_json::json!({
+///     "code": 1000,
+///     "message": "my error",
+/// }));
+/// # });
 /// ```
 ///
 /// # Downcast the error to concrete error type
-///
 /// ```
 /// use poem::{error::NotFoundError, Error};
 ///
@@ -94,10 +181,17 @@ impl Debug for ErrorSource {
 /// assert!(err.is::<NotFoundError>());
 /// assert_eq!(err.downcast_ref::<NotFoundError>(), Some(&NotFoundError));
 /// ```
-#[derive(Debug)]
 pub struct Error {
-    status: StatusCode,
+    as_response: AsResponse,
     source: ErrorSource,
+}
+
+impl Debug for Error {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Error")
+            .field("source", &self.source)
+            .finish()
+    }
 }
 
 impl Display for Error {
@@ -118,24 +212,23 @@ impl From<Infallible> for Error {
 
 impl<T: ResponseError + StdError + Send + Sync + 'static> From<T> for Error {
     fn from(err: T) -> Self {
-        let status = err.status();
-        Error::new(err, status)
+        Error {
+            as_response: AsResponse::from_type::<T>(),
+            source: ErrorSource::BoxedError(Box::new(err)),
+        }
     }
 }
 
 impl From<Box<dyn StdError + Send + Sync>> for Error {
     fn from(err: Box<dyn StdError + Send + Sync>) -> Self {
-        Error {
-            status: StatusCode::INTERNAL_SERVER_ERROR,
-            source: ErrorSource::BoxedError(err),
-        }
+        (StatusCode::INTERNAL_SERVER_ERROR, err).into()
     }
 }
 
 impl From<(StatusCode, Box<dyn StdError + Send + Sync>)> for Error {
     fn from((status, err): (StatusCode, Box<dyn StdError + Send + Sync>)) -> Self {
         Error {
-            status,
+            as_response: AsResponse::from_status(status),
             source: ErrorSource::BoxedError(err),
         }
     }
@@ -145,7 +238,7 @@ impl From<(StatusCode, Box<dyn StdError + Send + Sync>)> for Error {
 impl From<anyhow::Error> for Error {
     fn from(err: anyhow::Error) -> Self {
         Error {
-            status: StatusCode::INTERNAL_SERVER_ERROR,
+            as_response: AsResponse::from_status(StatusCode::INTERNAL_SERVER_ERROR),
             source: ErrorSource::Anyhow(err),
         }
     }
@@ -155,7 +248,7 @@ impl From<anyhow::Error> for Error {
 impl From<(StatusCode, anyhow::Error)> for Error {
     fn from((status, err): (StatusCode, anyhow::Error)) -> Self {
         Error {
-            status,
+            as_response: AsResponse::from_status(status),
             source: ErrorSource::Anyhow(err),
         }
     }
@@ -173,7 +266,7 @@ impl Error {
     #[inline]
     pub fn new<T: StdError + Send + Sync + 'static>(err: T, status: StatusCode) -> Self {
         Self {
-            status,
+            as_response: AsResponse::from_status(status),
             source: ErrorSource::BoxedError(Box::new(err)),
         }
     }
@@ -203,12 +296,6 @@ impl Error {
         Self::new(StringError(msg.into()), status)
     }
 
-    /// Returns the status code of this error.
-    #[inline]
-    pub fn status(&self) -> StatusCode {
-        self.status
-    }
-
     /// Downcast this error object by reference.
     #[inline]
     pub fn downcast_ref<T: StdError + Send + Sync + 'static>(&self) -> Option<&T> {
@@ -222,13 +309,13 @@ impl Error {
     /// Attempts to downcast the error to a concrete error type.
     #[inline]
     pub fn downcast<T: StdError + Send + Sync + 'static>(self) -> Result<T, Error> {
-        let status = self.status;
+        let as_response = self.as_response;
 
         match self.source {
             ErrorSource::BoxedError(err) => match err.downcast::<T>() {
                 Ok(err) => Ok(*err),
                 Err(err) => Err(Error {
-                    status,
+                    as_response,
                     source: ErrorSource::BoxedError(err),
                 }),
             },
@@ -236,7 +323,7 @@ impl Error {
             ErrorSource::Anyhow(err) => match err.downcast::<T>() {
                 Ok(err) => Ok(err),
                 Err(err) => Err(Error {
-                    status,
+                    as_response,
                     source: ErrorSource::Anyhow(err),
                 }),
             },
@@ -255,12 +342,7 @@ impl Error {
 
     /// Consumes this to return a response object.
     pub fn as_response(&self) -> Response {
-        let msg = match &self.source {
-            ErrorSource::BoxedError(err) => err.to_string(),
-            #[cfg(feature = "anyhow")]
-            ErrorSource::Anyhow(err) => err.to_string(),
-        };
-        Response::builder().status(self.status).body(msg)
+        self.as_response.as_response(self)
     }
 }
 
@@ -666,6 +748,17 @@ impl ResponseError for StaticFileError {
             StaticFileError::Io(_) => StatusCode::INTERNAL_SERVER_ERROR,
         }
     }
+
+    fn as_response(&self) -> Response {
+        let mut resp = Response::builder()
+            .status(self.status())
+            .body(self.to_string());
+        if let StaticFileError::RangeNotSatisfiable { size } = self {
+            resp.headers_mut()
+                .typed_insert(ContentRange::unsatisfied_bytes(*size));
+        }
+        resp
+    }
 }
 
 /// A possible error value occurred in the `SizeLimit` middleware.
@@ -779,12 +872,12 @@ mod tests {
             IoError::new(ErrorKind::AlreadyExists, "aaa"),
             StatusCode::BAD_GATEWAY,
         );
-        assert_eq!(err.status(), StatusCode::BAD_GATEWAY);
         assert!(err.is::<IoError>());
         assert_eq!(
             err.downcast_ref::<IoError>().unwrap().kind(),
             ErrorKind::AlreadyExists
         );
+        assert_eq!(err.as_response().status(), StatusCode::BAD_GATEWAY);
     }
 
     #[test]
@@ -792,12 +885,12 @@ mod tests {
         let boxed_err: Box<dyn StdError + Send + Sync> =
             Box::new(IoError::new(ErrorKind::AlreadyExists, "aaa"));
         let err: Error = Error::from((StatusCode::BAD_GATEWAY, boxed_err));
-        assert_eq!(err.status(), StatusCode::BAD_GATEWAY);
         assert!(err.is::<IoError>());
         assert_eq!(
             err.downcast_ref::<IoError>().unwrap().kind(),
             ErrorKind::AlreadyExists
         );
+        assert_eq!(err.as_response().status(), StatusCode::BAD_GATEWAY);
     }
 
     #[cfg(feature = "anyhow")]
@@ -805,11 +898,39 @@ mod tests {
     fn test_anyhow_error() {
         let anyhow_err: anyhow::Error = IoError::new(ErrorKind::AlreadyExists, "aaa").into();
         let err: Error = Error::from((StatusCode::BAD_GATEWAY, anyhow_err));
-        assert_eq!(err.status(), StatusCode::BAD_GATEWAY);
         assert!(err.is::<IoError>());
         assert_eq!(
             err.downcast_ref::<IoError>().unwrap().kind(),
             ErrorKind::AlreadyExists
+        );
+        assert_eq!(err.as_response().status(), StatusCode::BAD_GATEWAY);
+    }
+
+    #[tokio::test]
+    async fn test_custom_as_response() {
+        #[derive(Debug, thiserror::Error)]
+        #[error("my error")]
+        struct MyError;
+
+        impl ResponseError for MyError {
+            fn status(&self) -> StatusCode {
+                StatusCode::BAD_GATEWAY
+            }
+
+            fn as_response(&self) -> Response {
+                Response::builder()
+                    .status(self.status())
+                    .body("my error message")
+            }
+        }
+
+        let err = Error::from(MyError);
+        let resp = err.as_response();
+
+        assert_eq!(resp.status(), StatusCode::BAD_GATEWAY);
+        assert_eq!(
+            resp.into_body().into_string().await.unwrap(),
+            "my error message"
         );
     }
 }